feat: chemical maps with simd, ant grid sensor and first successful train w/ following food scent and interacting with it

+++ Assets/_Project/Scripts/Holdable System/Equipment/Pickaxe.cs
@@ -9,7 +9,7 @@ namespace Core.HoldableSystem
 
         protected override void UseBehavior(Vector2 worldPosition)
         {
-            _gridService.DamageTileAt(worldPosition, (int)Attributes.MiningStrength);
+            _gridService.DamageTileAt(worldPosition, Attributes.MiningStrength);
         }
     }
 }
